// TightLinesAI/lib/pierCastWindGrid.generated.ts
export type PierCastWindGridNode = {
  nodeId: string;
  lakeId: PierCastGreatLakeId;
  latitude: number;
  longitude: number;
};